The many and sep_by family of parsers use FromIterator to collect the values
By using FromIterator these kinds of parsers can turn their results into any sort of container instead of only a Vec. Most noticeably the chars and chars1 can be replaced by many and many1 parsers respectively.

The downside of this is that additional type annotations are required when using these parsers in case that the compiler is unable to infer the right type.