Add an extra method to Parser which can delay the generation of errors
Parsers such as `choice` and `.or` can result in the creation of several errors before they find the correct parser. The `parse_ok` and `add_error` methods are added to allow parsers to avoid allocating a vector to store the error in some cases. Instead the parser promises that it can generate the error at a later point if asked for it through the `add_error` method.

This gives a 30% speedup on the json parser benchmark but since it mostly contained strings for values it was already optimized to generate less by placing the string parser first in the `.or` chain it can be expected to see larger improvements for more balanced code.
(placing the string parser last showed a ~90% improvement)