feat: Add the StreamOnce trait
This allows `BufferedStream` to be bounded by `StreamOnce` instead of
`Iterator` allowing the inner stream to indicate other errors than end of
input. `Stream` should still be used as a trait bound in parser code but
`StreamOnce` must now be used when implementing a stream.

BREAKING CHANGE: When implementing a `Stream` `StreamOnce` must now be
used.