Add map functions for Error<> and Info<> ranges. (#86)
It's possible to `map_err_range` for `ParseResult<>` too, but it's
awkward because the output type needs to be a compatible `StreamOnce`.
As suggested in
https://github.com/Marwes/combine/issues/86#issuecomment-280487165,
it's probably best to either change the parse result type entirely, or
wait for https://github.com/rust-lang/rust/issues/21903.

This at least helps consumers convert `ParseError<>` into something
that can implement `std::fmt::Display`.