fix: Make the positions of slice streams harder to misuse
Previously these streams had a `usize` which was really a pointer offset
as their position which easily led to user errors. By wrapping the
offset in a newtype we get custom formatting which should make it
clearer that something strange is going on. As well as a simple way of
converting the offset into a position.

Fixes #104

BREAKING CHANGE

Change `usize` to `PointerOffset` where applicable.
Change `translate_position` calls into the same method on `PointerOffset`