refactor: Split the position out of the StreamOnce trait
The `StreamOnce` trait do not need the position to work even if many users of the trait needs it. So this way it becomes possible for wrappers such as `State` to carry types implementing `StreamOnce` but which are not `Positioned`

BREAKING CHANGE

Implementors of `StreamOnce` should implement `Positioned` as well.
`Positioned` must be imported if the `position` methoid is used