fix: not_followed_by now does not consume any input
The `not_followed_by` doc says that it never consumes any input,
however it does if parser succeeds.

The implementation was missing a `try` before which made the parser preserve the `Consumed` status if `P` consumed input before failing. This is because `try` only makes errors look like they haven't consumed any input and at the point of the inner `try` the parse status is still `Ok`.