feat: Don't require `Clone` for `Stream`
Instead we require the `Resetable` which lets `Stream` types potentially
implement a cheaper way to reset themselves to an earlier position. If
they implement `Clone` there is a macro to trivially generate the
implementation

BREAKING CHANGE

If you have implemented a custom `Stream` type you need to implement
`Resetable`. The `clone_resetable` can be used to do so easily.