feat: Don't have the error type be generic for uncons*
This was intended as a way to let `StreamOnce` instances to be wrapped
and let the wrapping instance decide which error type that were
returned. However this isn't necessary except for rare cases such as
when a `BufferedStream` wraps a stream with poor errors and is wrapped
with a stream providing good errors (see the fix in
`tests/buffered_stream` and even then it may be fixed by simply moving
the stream wrapper with good error inside the `BufferedStream` instance.

All in all the generic solution is more confusing than its worth so
removing it seems prudent. If it becomes necessary it can always be
added back backwards compatibly as a default method.