Unification of style
For existing corrections, I unified the style of specifying links
in the places adjacent to them.
(Not all links are unified.)

It also doubles as a check to see if Travis CI still fails.
If the previous error log is true, that may be a bug in rustc,
but I haven't investigated the smallest code to reproduce it yet.
(Maybe rust-lang/rust 79560 is related?)