feat: Add decode_tokio
This works mostly the same as the regular `decode` function but it
can be plugged into both `decode` and `decode_eof` it will work as long
as the `eof` variant marks the input as non-partial.

Also ensures that unexpected end of inputs always get returned so that
we can detect them at the real end of input and return `None`.

The reqular `decode` function ended up still working as long as the
stream was always reported as partial as the parsers would endeavour
to return unexpected end of input in those cases.